TIDIJ-72 Created UdpManager class for handling udp packets and better code organisation
